requirements.txt
streamlit
pandas
gspread
google-auth
google-auth-httplib2
google-auth-oauthlib
google-genai
yfinance
plotly
openpyxl
requests
Pillow
